test(scripts): SMI-5141 split oversized release/audit test files under 500-line gate
Split prepare-release.test.ts (648) and audit-standards.test.ts (581) into
concern-scoped siblings, each well under the check-file-length 500-line gate,
and clear the prepare-release.test.ts grandfather entry from
scripts/check-file-length.ignore.

- audit-standards.test.ts (313): parseSemver / satisfies / extractCompletionIssues / Check 23
- audit-standards-export-drift.test.ts (new): parseTsExports / collectTsEntryExports / extractSmokeTestRequiredArrays
- audit-standards-publish-gate.test.ts: + auditPublishYmlDependentGate (relocated, co-located with the related required-gate suite)
- prepare-release.test.ts (236): config + version-sync + updateWorkspaceDependencies
- prepare-release-collision.test.ts (new): fetchNpmLatest + checkVersionCollision matrix + fail-closed + fixture parity
- prepare-release-reserved-range.test.ts (new): reserved version-range guard

No coverage change: 119 tests across the 7 files, same assertions redistributed.
