+++ scripts/tests/audit-standards.test.ts
@@ -2,14 +2,20 @@
  * Tests for the pure helpers in scripts/audit-standards-helpers.mjs.
  *
  * Covers SMI-3987 (Check 11 dedup awareness via `satisfies`) and Check 23
- * cite-in-body filtering (via `extractCompletionIssues`), plus SMI-3986
- * (Check 23 worktree-aware `git rev-parse --git-common-dir` resolution).
+ * cite-in-body filtering (via `extractCompletionIssues`).
  *
  * The pure helpers are imported via dynamic ESM import from a small
  * companion file, matching the convention used by
- * scripts/tests/check-supply-chain-pins.test.ts. The integration test for
- * `gitCommonDir` uses a real tmpdir-based git repo + worktree (the only
- * way to validate the SMI-3986 fix end-to-end).
+ * scripts/tests/check-supply-chain-pins.test.ts.
+ *
+ * Sibling files (split out under SMI-5141 to keep this file under the 500-line
+ * CI gate):
+ *   - audit-standards-export-drift.test.ts — parseTsExports / collectTsEntryExports
+ *     / extractSmokeTestRequiredArrays (SMI-4193 Check 29).
+ *   - audit-standards-publish-gate.test.ts — auditPublishYmlDependentGate
+ *     (SMI-5060/5066) + auditPublishYmlRequiredGates (SMI-5123).
+ *   - audit-standards-gitcommondir.test.ts — SMI-3986 worktree
+ *     `git rev-parse --git-common-dir` integration.
  */
 import { describe, expect, it } from 'vitest'
 import { readFileSync } from 'node:fs'
@@ -21,30 +27,9 @@ const helpers = (await import('../audit-standards-helpers.mjs')) as {
   parseSemver: (v: string) => [number, number, number] | null
   satisfies: (version: string, spec: string) => boolean
   extractCompletionIssues: (subject: string, body: string) => Set<string>
-  parseTsExports: (content: string) => { names: Set<string>; starFrom: string[] }
-  collectTsEntryExports: (
-    entryPath: string,
-    readFile: (p: string) => string | null,
-    resolveModule: (fromFile: string, spec: string) => string | null
-  ) => Set<string>
-  extractSmokeTestRequiredArrays: (content: string) => { name: string; arrayIndex: number }[]
-  auditPublishYmlDependentGate: (content: string) => {
-    matches: Array<{ lineno: number; line: string; pkg: string; outputKey: string }>
-    failures: Array<{ lineno: number; line: string; pkg: string; outputKey: string }>
-  }
-  PUBLISH_JOB_TO_OUTPUT_ALIAS: Record<string, string>
 }
 
-const {
-  parseSemver,
-  satisfies,
-  extractCompletionIssues,
-  parseTsExports,
-  collectTsEntryExports,
-  extractSmokeTestRequiredArrays,
-  auditPublishYmlDependentGate,
-  PUBLISH_JOB_TO_OUTPUT_ALIAS,
-} = helpers
+const { parseSemver, satisfies, extractCompletionIssues } = helpers
 
 const __dirname = dirname(fileURLToPath(import.meta.url))
 
@@ -321,261 +306,8 @@ describe('audit-standards Check 23: NON_SOURCE_PREFIXES (conventional commit typ
   })
 })
 
-// SMI-3986 gitCommonDir worktree integration test extracted to
-// scripts/tests/audit-standards-gitcommondir.test.ts to keep this file under
-// the 500-line CI gate (SMI-3493) after the SMI-4693 helper migration.
-
-// ============================================================================
-// SMI-4193: Smoke-test export drift (Check 29)
-// ============================================================================
-
-describe('parseTsExports', () => {
-  it('extracts named exports from export { ... }', () => {
-    const src = `export { Foo, Bar as Baz, type Qux } from './x.js'\nexport { Alone }`
-    const { names, starFrom } = parseTsExports(src)
-    expect([...names].sort()).toEqual(['Alone', 'Baz', 'Foo', 'Qux'])
-    expect(starFrom).toEqual([])
-  })
-
-  it('extracts function/const/class/enum/interface/type declarations', () => {
-    const src = `
-      export function foo() {}
-      export async function asyncFoo() {}
-      export const bar = 1
-      export class Baz {}
-      export enum MyEnum {}
-      export interface MyIface {}
-      export type MyType = string
-    `
-    const { names } = parseTsExports(src)
-    expect([...names].sort()).toEqual(
-      ['MyEnum', 'MyIface', 'MyType', 'asyncFoo', 'bar', 'Baz', 'foo'].sort()
-    )
-  })
-
-  it('records export * from chains in starFrom, not names', () => {
-    const src = `export * from './exports/services.js'\nexport * from './exports/repositories.js'`
-    const { names, starFrom } = parseTsExports(src)
-    expect(names.size).toBe(0)
-    expect(starFrom).toEqual(['./exports/services.js', './exports/repositories.js'])
-  })
-
-  it('ignores exports inside block comments (SMI-4189 pattern)', () => {
-    const src = `/* export { Removed } from './old.js' */\nexport { Kept }`
-    const { names } = parseTsExports(src)
-    expect([...names]).toEqual(['Kept'])
-  })
-})
-
-describe('collectTsEntryExports', () => {
-  it('walks export * chains and unions all names', () => {
-    const files: Record<string, string> = {
-      '/src/index.ts': `export * from './barrel.js'\nexport { Direct }`,
-      '/src/barrel.ts': `export { Nested1, Nested2 }`,
-    }
-    const readFile = (p: string) => files[p] ?? null
-    const resolveModule = (from: string, spec: string) => {
-      if (from === '/src/index.ts' && spec === './barrel.js') return '/src/barrel.ts'
-      return null
-    }
-    const result = collectTsEntryExports('/src/index.ts', readFile, resolveModule)
-    expect([...result].sort()).toEqual(['Direct', 'Nested1', 'Nested2'])
-  })
-
-  it('tolerates unresolvable barrels without throwing', () => {
-    const files: Record<string, string> = {
-      '/src/index.ts': `export * from './missing.js'\nexport { Kept }`,
-    }
-    const readFile = (p: string) => files[p] ?? null
-    const resolveModule = () => null
-    const result = collectTsEntryExports('/src/index.ts', readFile, resolveModule)
-    expect([...result]).toEqual(['Kept'])
-  })
-
-  it('guards against circular export * chains', () => {
-    const files: Record<string, string> = {
-      '/a.ts': `export * from './b.js'\nexport { FromA }`,
-      '/b.ts': `export * from './a.js'\nexport { FromB }`,
-    }
-    const readFile = (p: string) => files[p] ?? null
-    const resolveModule = (_from: string, spec: string) =>
-      spec === './a.js' ? '/a.ts' : spec === './b.js' ? '/b.ts' : null
-    const result = collectTsEntryExports('/a.ts', readFile, resolveModule)
-    expect([...result].sort()).toEqual(['FromA', 'FromB'])
-  })
-})
-
-describe('extractSmokeTestRequiredArrays', () => {
-  it('captures names from every required array with stable arrayIndex', () => {
-    const src = `
-      const required = ['A', 'B', 'C']
-      // later
-      const required = [
-        'D',
-        'E'
-      ]
-    `
-    const out = extractSmokeTestRequiredArrays(src)
-    expect(out).toEqual([
-      { name: 'A', arrayIndex: 0 },
-      { name: 'B', arrayIndex: 0 },
-      { name: 'C', arrayIndex: 0 },
-      { name: 'D', arrayIndex: 1 },
-      { name: 'E', arrayIndex: 1 },
-    ])
-  })
-
-  it('ignores required arrays that appear only inside comments', () => {
-    const src = `// const required = ['Commented']\nconst required = ['Real']`
-    const out = extractSmokeTestRequiredArrays(src)
-    expect(out).toEqual([{ name: 'Real', arrayIndex: 0 }])
-  })
-
-  it('returns empty array when no required declarations exist', () => {
-    const src = `const something = ['NotRequired']`
-    expect(extractSmokeTestRequiredArrays(src)).toEqual([])
-  })
-
-  it('detects SMI-4189 regression: CategoryRepository drift against a mock core', () => {
-    // Simulates the exact pattern that shipped in smoke-test@0.4.4 and caused
-    // the failed republish. The `required` array references `CategoryRepository`,
-    // but the mock core only exports `SkillRepository`.
-    const smokeSrc = `const required = ['SkillRepository', 'CategoryRepository', 'createDatabaseSync']`
-    const files: Record<string, string> = {
-      '/core/index.ts': `export { SkillRepository, createDatabaseSync }`,
-    }
-    const coreExports = collectTsEntryExports(
-      '/core/index.ts',
-      (p) => files[p] ?? null,
-      () => null
-    )
-    const entries = extractSmokeTestRequiredArrays(smokeSrc)
-    const missing = entries.filter((e) => !coreExports.has(e.name))
-    expect(missing.map((m) => m.name)).toEqual(['CategoryRepository'])
-  })
-})
-
-/**
- * SMI-5066: regression tests for the generalized Check 48 helper. Background:
- * SMI-5060 introduced the paired-predicate invariant (every
- * `needs.publish-<pkg>.result == 'skipped'` clause must be guarded by
- * `pre-publish-check.outputs.<outputKey>-exists == 'true'` within ±1 line).
- * SMI-5066 generalized it from publish-core-only to any publish-<pkg>, with an
- * alias map for the pre-existing `publish-mcp-server` → `mcp-exists` outlier.
- *
- * These tests guard against future regressions in either the regex or the
- * alias-map convention.
- */
-describe('auditPublishYmlDependentGate (SMI-5060/SMI-5066)', () => {
-  it('PUBLISH_JOB_TO_OUTPUT_ALIAS maps mcp-server to mcp', () => {
-    // The alias map exists to document and handle the convention drift
-    // between the publish-* job name and the pre-publish-check output key.
-    expect(PUBLISH_JOB_TO_OUTPUT_ALIAS['mcp-server']).toBe('mcp')
-  })
-
-  it('passes when publish-core skipped clause is paired with core-exists guard', () => {
-    const yml = [
-      'jobs:',
-      '  publish-mcp-server:',
-      '    if: |',
-      "      (needs.publish-core.result == 'success' ||",
-      "       (needs.publish-core.result == 'skipped' && needs.pre-publish-check.outputs.core-exists == 'true'))",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(1)
-    expect(matches[0].pkg).toBe('core')
-    expect(matches[0].outputKey).toBe('core')
-    expect(failures).toHaveLength(0)
-  })
-
-  it('passes when publish-enterprise skipped clause is paired with enterprise-exists guard', () => {
-    const yml = [
-      'jobs:',
-      '  publish-enterprise:',
-      '    if: |',
-      "      (needs.publish-core.result == 'success' ||",
-      "       (needs.publish-core.result == 'skipped' && needs.pre-publish-check.outputs.core-exists == 'true'))",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(1)
-    expect(matches[0].pkg).toBe('core')
-    expect(matches[0].outputKey).toBe('core')
-    expect(failures).toHaveLength(0)
-  })
-
-  it('passes when publish-mcp-server skipped clause is paired with mcp-exists guard (alias map)', () => {
-    const yml = [
-      'jobs:',
-      '  publish-cli:',
-      '    if: |',
-      "      (needs.publish-mcp-server.result == 'success' ||",
-      "       (needs.publish-mcp-server.result == 'skipped' && needs.pre-publish-check.outputs.mcp-exists == 'true'))",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(1)
-    expect(matches[0].pkg).toBe('mcp-server')
-    expect(matches[0].outputKey).toBe('mcp') // alias resolved
-    expect(failures).toHaveLength(0)
-  })
-
-  it('fails when publish-core skipped clause has no paired guard', () => {
-    const yml = [
-      'jobs:',
-      '  publish-mcp-server:',
-      '    if: |',
-      "      (needs.publish-core.result == 'success' ||",
-      "       needs.publish-core.result == 'skipped')",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(1)
-    expect(failures).toHaveLength(1)
-    expect(failures[0].pkg).toBe('core')
-    expect(failures[0].outputKey).toBe('core')
-  })
-
-  it('fails when publish-enterprise skipped clause has no paired guard', () => {
-    const yml = [
-      'jobs:',
-      '  publish-enterprise:',
-      '    if: |',
-      "      (needs.publish-core.result == 'success' ||",
-      "       needs.publish-core.result == 'skipped')",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(1)
-    expect(failures).toHaveLength(1)
-    expect(failures[0].pkg).toBe('core')
-  })
-
-  it('ignores comment-only lines containing the skipped phrase', () => {
-    const yml = [
-      'jobs:',
-      "  # Documentation: when needs.publish-core.result == 'skipped' fires,",
-      '  # the guard handles it.',
-      '  publish-mcp-server:',
-      "    if: needs.pre-publish-check.outputs.mcp-exists != 'true'",
-    ].join('\n')
-    const { matches, failures } = auditPublishYmlDependentGate(yml)
-    expect(matches).toHaveLength(0)
-    expect(failures).toHaveLength(0)
-  })
-
-  it('fails when paired guard is more than ±1 line away from the skipped clause', () => {
-    // The window is intentionally tight (±1 line) to match the canonical YAML
-    // shape produced by gh actions multi-line `if: |` blocks. A guard that
-    // drifts farther is a smell — either the YAML restructured or the guard
-    // was decoupled.
-    const yml = [
-      'jobs:',
-      '  publish-mcp-server:',
-      '    if: |',
-      "      (needs.publish-core.result == 'success' ||",
-      "       needs.publish-core.result == 'skipped') &&",
-      '      true &&',
-      '      true &&',
-      "      needs.pre-publish-check.outputs.core-exists == 'true'",
-    ].join('\n')
-    const { failures } = auditPublishYmlDependentGate(yml)
-    expect(failures).toHaveLength(1)
-  })
-})
+// SMI-3986 gitCommonDir worktree integration test lives in
+// scripts/tests/audit-standards-gitcommondir.test.ts. Smoke-test export-drift
+// (Check 29) and publish.yml gating helpers were split into
+// audit-standards-export-drift.test.ts and audit-standards-publish-gate.test.ts
+// respectively (SMI-5141) to keep this file under the 500-line CI gate.
